+++ src/Staff.php
@@ -102,7 +102,7 @@ class Staff extends Plugin
     /**
      * @var Settings
      */
-    public static $settings;
+    public static Settings $settings;
 
     /**
      * @var View
@@ -117,7 +117,7 @@ class Staff extends Plugin
      *
      * @var string
      */
-    public $schemaVersion = '1.0.1';
+    public $schemaVersion = '1.0.2';
 
     /**
      * Set to `true` if the plugin should have its own section (main nav item) in the control panel.
@@ -287,11 +287,9 @@ class Staff extends Plugin
             ];
         }
 
-        $navItem = array_merge($navItem, [
+        return array_merge($navItem, [
             'subnav' => $subNavs,
         ]);
-
-        return $navItem;
     }
 
     // Protected Methods