+++ src/gql/queries/PayRunEntry.php
@@ -20,14 +20,14 @@ class PayRunEntry extends Query
         }
 
         return [
-            'PayRunEntries' => [
+            'PayrunEntries' => [
                 'type' => Type::listOf(PayRunEntryInterface::getType()),
                 'args' => PayRunEntryArguments::getArguments(),
                 'resolve' => PayRunEntryResolver::class . '::resolve',
                 'description' => 'This query is used to query for all the Payrun Entries.',
                 'complexity' => GqlHelper::relatedArgumentComplexity(),
             ],
-            'PayRunEntry' => [
+            'PayrunEntry' => [
                 'type' => PayRunEntryInterface::getType(),
                 'args' => PayRunEntryArguments::getArguments(),
                 'resolve' => PayRunEntryResolver::class . '::resolveOne',